pyproject.toml
[tool.poetry]
name = "pytesting"
version = "0.1.0"
description = ""
authors = ["Fasih Khatib <hellofasih.confound928@passinbox.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.3.3"
hypothesis = "^6.119.3"
peewee = "^3.17.8"
psycopg2-binary = "^2.9.10"
flask = "^3.1.0"
python-dateutil = "^2.9.0.post0"
gunicorn = "^23.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
